Ease setup of new boards by dynamically creating/identifying lists and labels
Previously we depended on configuration-defined IDs for Trello labels and lists.  This implied that these labels and lists would have to be manually configured in Trello, and then their IDs tediously found and replicated to the config file.  This commit adds a new model, TrelloList, that will track list IDs in the database instead of the config file.  The config file has been updated to include a set of default captions and symbolic names for both Trello lists and labels.  Two additional CLI commands have been added (`initlists`, `initlabels`, and `initboard` which is a convenience command that calls the first two commands).  These commands will automatically configure a Trello board with the appropriate lists and labels, and store the relevant IDs to the database.  The commands are built to be idempotent, so they can be run against a new Trello board, an existing Trello board, or a copied board (which would have the correct lists, but different IDs).