Add support for sending emails using more than just the submission status
The `send_email` task now takes a template name to decide which email to send.  For existing state-change emails, we use the same logic that was already present to convert from the current state to the correct email template and subject line.  In addition to the template name, `send_email` also now supports variable template parameters.  It's up to the caller to understand which template parameters are valid for the template they're requesting.